Module brownstone::builder

A low level builder type for creating fixed size arrays. See ArrayBuilder for details.

Structs

ArrayBuilder

Low-level builder type for [T; N] arrays. Uses a push + finish interface to build an array 1 element at a time.

Overflow

Error type returned from ArrayBuilder::try_push, indicating that the builder was already full. Includes the value that couldn’t be pushed to the array.

Enums

PushResult

Result type returned from ArrayBuilder::push, indicating whether the array is full after the push. ArrayBuilder::push panics on overflow, so this only indicates if there is room to push additional elements.