brownstone::builder
Struct ArrayBuilder

pub struct ArrayBuilder<T, const N: usize> { /* private fields */ }

Low-level builder type for [T; N] arrays. Uses a push + finish interface to build an array 1 element at a time.

The interface provided by this type is fairly low level; most of its methods are fallible in some way (returning a Result or panicking on errors). Consider instead the misuse-resistant move_builder::ArrayBuilder, which uses ownership semantics to provide only infallible operations, or the build! macro at the top level of the crate.

Implementations

impl<T, const N: usize> ArrayBuilder<T, N>

pub const fn new() -> Self

Create a new, empty ArrayBuilder.

pub fn is_full(&self) -> bool

Returns true if every element in the array is initialized. If the builder is full, the next call to finish will return the built array.

pub fn is_empty(&self) -> bool

Returns true if no elements in the array are initialized.

pub fn len(&self) -> usize

Returns the number of initialized elements in the array.

pub unsafe fn push_unchecked(&mut self, value: T) -> PushResult

Add an initialized element to the array, without performing a bounds check.

Safety

This must only be called when the builder is not full.

pub fn try_push(&mut self, value: T) -> Result<PushResult, Overflow<T>>

Try to add an initialized element to the array. Returns an error if the array is already full, or a PushResult indicating if the array is now full and can be retrieved via finish.

pub fn push(&mut self, value: T) -> PushResult

Add an initialized element to the array. Returns a PushResult indicating if the array is now full and can be retrieved via finish.

Panics

Panics if the array is already full.

pub unsafe fn finish_unchecked(self) -> [T; N]

Return the fully initialized array without checking that itโs fully initialized.

Safety

This must only be called when the builder is full.

pub fn try_finish(self) -> Result<[T; N], Self>

Try to return the fully initialized array. Returns the builder if the array isnโt fully initialized yet.

pub fn finish(self) -> [T; N]

Return the fully initialized array.

Panics

Panics if the array isnโt fully initialized yet.

pub fn finished_slice(&self) -> &[T]

Get the slice of the array that has already been initialized.

pub fn finished_slice_mut(&mut self) -> &mut [T]

Get the mutable slice of the array that has already been initialized.