Crate brownstone

brownstone is a library for building fixed-size arrays. It provides build, a macro that builds an array by evaluating an expression once for each element in the array. It also provides a low-level builder type, with a push + finish interface, as well as a misuse-immune builder type with a move-based interface that can never panic or return errors.

Modules

builder

A low level builder type for creating fixed size arrays. See ArrayBuilder for details.

move_builder

A misuse-immune array builder. See ArrayBuilder for details and examples.

Macros

build

Build an array with an expression.